clean up

+++ raven_back/lib/models/account.dart
@@ -1,7 +1,6 @@
 import 'package:equatable/equatable.dart';
 import 'package:raven/models/balance.dart';
 import 'package:raven/records.dart' as records;
-import 'package:raven/utils/rate.dart';
 
 // ignore: must_be_immutable
 class Account extends Equatable {
@@ -36,13 +35,12 @@ class Account extends Equatable {
     var rvn = getTotalRVN();
     var usd;
     if (rvn.value > 0) {
-      /// we should alrady know this... and then only updated it if we need it
-      //var rate = await conversionRate('usd');
       /// could get it from the conversion rate reservoir - should access through service?
       //import 'package:raven/reservoirs.dart';
       var rate = rates.getRvnToUsd();
       usd = BalanceUSD(
-          confirmed: rvn.confirmed * rate, unconfirmed: rvn.confirmed * rate);
+          confirmed: (rvn.confirmed * rate).toDouble(),
+          unconfirmed: (rvn.confirmed * rate).toDouble());
     }
     return usd;
   }
